团购
分类、城市、排序

+++ XC/Classes/Deal(团购)/View/XUHomeDropdown.h
@@ -19,31 +19,31 @@
  *  左边表格每一行的标题
  *  @param row          行号
  */
-- (NSString *)homeDropdown:(XUHomeDropdown *)homeDropdown titleForRowInMainTable:(int)row;
+- (NSString *)homeDropdown:(XUHomeDropdown *)homeDropdown titleForRowInMainTable:(NSInteger)row;
 /**
  *  左边表格每一行的子数据
  *  @param row          行号
  */
-- (NSArray *)homeDropdown:(XUHomeDropdown *)homeDropdown subdataForRowInMainTable:(int)row;
+- (NSArray *)homeDropdown:(XUHomeDropdown *)homeDropdown subdataForRowInMainTable:(NSInteger)row;
 
 @optional
 /**
  *  左边表格每一行的图标
  *  @param row          行号
  */
-- (NSString *)homeDropdown:(XUHomeDropdown *)homeDropdown iconForRowInMainTable:(int)row;
+- (NSString *)homeDropdown:(XUHomeDropdown *)homeDropdown iconForRowInMainTable:(NSInteger)row;
 /**
  *  左边表格每一行的选中图标
  *  @param row          行号
  */
-- (NSString *)homeDropdown:(XUHomeDropdown *)homeDropdown selectedIconForRowInMainTable:(int)row;
+- (NSString *)homeDropdown:(XUHomeDropdown *)homeDropdown selectedIconForRowInMainTable:(NSInteger)row;
 @end
 
 @protocol XUHomeDropdownDelegate <NSObject>
 
 @optional
-- (void)homeDropdown:(XUHomeDropdown *)homeDropdown didSelectRowInMainTable:(int)row;
-- (void)homeDropdown:(XUHomeDropdown *)homeDropdown didSelectRowInSubTable:(int)subrow inMainTable:(int)mainRow;
+- (void)homeDropdown:(XUHomeDropdown *)homeDropdown didSelectRowInMainTable:(NSInteger)row;
+- (void)homeDropdown:(XUHomeDropdown *)homeDropdown didSelectRowInSubTable:(NSInteger)subrow inMainTable:(NSInteger)mainRow;
 @end
 
 @interface XUHomeDropdown : UIView
